Added a UserLogin class and tests
Added a class that stores user handle, hash, and salt.
Can be used in the compute layer during authentication.

Added a method to the UserConnector interface to get user login objects
and implemented it in DBUserConnector.

Added a test to ensure user login data can be retrieved using a handle.

Damien Robichaud & Ryan Lowe